+++ app/src/main/AndroidManifest.xml
@@ -190,7 +190,7 @@
         </provider>
 
         <service android:name=".application.ApplicationService" />
-        <service android:name=".service.intentservice.TransactionListService"
+        <service android:name=".service.intent.TransactionListService"
             android:permission="android.permission.BIND_JOB_SERVICE"
             android:exported="true"/>
