+++ app/src/main/java/org/nervos/neuron/activity/AboutUsActivity.java
@@ -6,7 +6,7 @@ import android.widget.TextView;
 import org.nervos.neuron.R;
 import org.nervos.neuron.view.SettingButtonView;
 import org.nervos.neuron.view.TitleBar;
-import org.nervos.neuron.service.httpservice.HttpUrls;
+import org.nervos.neuron.service.http.HttpUrls;
 
 /**
  * Created by 包俊 on 2018/7/30.
