+++ app/src/main/java/org/nervos/neuron/activity/AddTokenActivity.java
@@ -16,9 +16,9 @@ import android.widget.Toast;
 import org.nervos.neuron.item.ChainItem;
 import org.nervos.neuron.item.TokenEntity;
 import org.nervos.neuron.item.WalletItem;
-import org.nervos.neuron.service.httpservice.EthRpcService;
-import org.nervos.neuron.service.httpservice.HttpUrls;
-import org.nervos.neuron.service.httpservice.AppChainRpcService;
+import org.nervos.neuron.service.http.EthRpcService;
+import org.nervos.neuron.service.http.HttpUrls;
+import org.nervos.neuron.service.http.AppChainRpcService;
 import org.nervos.neuron.R;
 import org.nervos.neuron.item.TokenItem;
 import org.nervos.neuron.util.AddressUtil;
