+++ app/src/main/java/org/nervos/neuron/activity/MainActivity.java
@@ -15,9 +15,9 @@ import org.nervos.neuron.fragment.AppFragment;
 import org.nervos.neuron.fragment.SettingsFragment;
 import org.nervos.neuron.fragment.wallet.view.WalletsFragment;
 import org.nervos.neuron.item.TokenItem;
-import org.nervos.neuron.service.httpservice.AppChainRpcService;
-import org.nervos.neuron.service.httpservice.HttpUrls;
-import org.nervos.neuron.service.intentservice.TransactionListService;
+import org.nervos.neuron.service.http.AppChainRpcService;
+import org.nervos.neuron.service.http.HttpUrls;
+import org.nervos.neuron.service.intent.TransactionListService;
 import org.nervos.neuron.util.ConstUtil;
 import org.nervos.neuron.util.qrcode.CodeUtils;
 import org.nervos.neuron.util.db.DBWalletUtil;
