+++ app/src/main/java/org/nervos/neuron/activity/PwdUnlockActivity.java
@@ -15,7 +15,7 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import org.nervos.neuron.R;
-import org.nervos.neuron.service.httpservice.WalletService;
+import org.nervos.neuron.service.http.WalletService;
 import org.nervos.neuron.view.SelectWalletPopupWindow;
 import org.nervos.neuron.item.WalletItem;
 import org.nervos.neuron.util.db.DBWalletUtil;
