+++ app/src/main/java/org/nervos/neuron/activity/TransactionDetailActivity.java
@@ -16,8 +16,8 @@ import com.yanzhenjie.permission.Permission;
 import org.nervos.neuron.R;
 import org.nervos.neuron.item.TransactionItem;
 import org.nervos.neuron.item.WalletItem;
-import org.nervos.neuron.service.httpservice.AppChainRpcService;
-import org.nervos.neuron.service.httpservice.NeuronSubscriber;
+import org.nervos.neuron.service.http.AppChainRpcService;
+import org.nervos.neuron.service.http.NeuronSubscriber;
 import org.nervos.neuron.util.ConstUtil;
 import org.nervos.neuron.util.NumberUtil;
 import org.nervos.neuron.util.SharePicUtils;
