+++ app/src/main/java/org/nervos/neuron/activity/transactionlist/model/TokenDescribeModel.java
@@ -5,8 +5,8 @@ import android.text.TextUtils;
 import com.google.gson.Gson;
 
 import org.nervos.neuron.item.EthErc20TokenInfoItem;
-import org.nervos.neuron.service.httpservice.HttpUrls;
-import org.nervos.neuron.service.httpservice.HttpService;
+import org.nervos.neuron.service.http.HttpUrls;
+import org.nervos.neuron.service.http.HttpService;
 import org.nervos.neuron.util.AddressUtil;
 import org.web3j.crypto.Keys;
