+++ app/src/main/java/org/nervos/neuron/activity/transactionlist/presenter/TransactionListPresenter.java
@@ -15,10 +15,10 @@ import org.nervos.neuron.item.ChainItem;
 import org.nervos.neuron.item.EthErc20TokenInfoItem;
 import org.nervos.neuron.item.TokenItem;
 import org.nervos.neuron.item.TransactionItem;
-import org.nervos.neuron.service.httpservice.AppChainTransactionService;
-import org.nervos.neuron.service.httpservice.HttpService;
-import org.nervos.neuron.service.httpservice.HttpUrls;
-import org.nervos.neuron.service.httpservice.TokenService;
+import org.nervos.neuron.service.http.AppChainTransactionService;
+import org.nervos.neuron.service.http.HttpService;
+import org.nervos.neuron.service.http.HttpUrls;
+import org.nervos.neuron.service.http.TokenService;
 import org.nervos.neuron.util.AddressUtil;
 import org.nervos.neuron.util.CurrencyUtil;
 import org.nervos.neuron.util.db.DBAppChainTransactionsUtil.TokenType;
