+++ app/src/main/java/org/nervos/neuron/activity/transfer/TransferActivity.java
@@ -23,8 +23,8 @@ import org.nervos.neuron.activity.NBaseActivity;
 import org.nervos.neuron.activity.QrCodeActivity;
 import org.nervos.neuron.item.TransactionInfo;
 import org.nervos.neuron.item.WalletItem;
-import org.nervos.neuron.service.httpservice.EthRpcService;
-import org.nervos.neuron.service.httpservice.WalletService;
+import org.nervos.neuron.service.http.EthRpcService;
+import org.nervos.neuron.service.http.WalletService;
 import org.nervos.neuron.util.AddressUtil;
 import org.nervos.neuron.util.Blockies;
 import org.nervos.neuron.util.NumberUtil;
