+++ app/src/main/java/org/nervos/neuron/activity/transfer/TransferPresenter.java
@@ -10,12 +10,12 @@ import org.nervos.neuron.item.CurrencyItem;
 import org.nervos.neuron.item.TokenItem;
 import org.nervos.neuron.item.TransactionInfo;
 import org.nervos.neuron.item.WalletItem;
-import org.nervos.neuron.service.httpservice.AppChainRpcService;
-import org.nervos.neuron.service.httpservice.EthRpcService;
-import org.nervos.neuron.service.httpservice.HttpUrls;
-import org.nervos.neuron.service.httpservice.NeuronSubscriber;
-import org.nervos.neuron.service.httpservice.TokenService;
-import org.nervos.neuron.service.httpservice.WalletService;
+import org.nervos.neuron.service.http.AppChainRpcService;
+import org.nervos.neuron.service.http.EthRpcService;
+import org.nervos.neuron.service.http.HttpUrls;
+import org.nervos.neuron.service.http.NeuronSubscriber;
+import org.nervos.neuron.service.http.TokenService;
+import org.nervos.neuron.service.http.WalletService;
 import org.nervos.neuron.util.ConstUtil;
 import org.nervos.neuron.util.CurrencyUtil;
 import org.nervos.neuron.util.NumberUtil;
