+++ app/src/main/java/org/nervos/neuron/service/intent/TransactionListService.java
@@ -1,10 +1,10 @@
-package org.nervos.neuron.service.intentservice;
+package org.nervos.neuron.service.intent;
 
 import android.content.Context;
 import android.content.Intent;
 import android.support.v4.app.JobIntentService;
 
-import org.nervos.neuron.service.httpservice.AppChainTransactionService;
+import org.nervos.neuron.service.http.AppChainTransactionService;
 
 /**
  * Created by BaojunCZ on 2018/10/11.
