+++ app/src/main/java/org/nervos/neuron/util/web/WebAppUtil.java
@@ -21,9 +21,9 @@ import org.nervos.neuron.event.AppCollectEvent;
 import org.nervos.neuron.event.AppHistoryEvent;
 import org.nervos.neuron.item.AppItem;
 import org.nervos.neuron.item.ChainItem;
-import org.nervos.neuron.service.httpservice.HttpUrls;
-import org.nervos.neuron.service.httpservice.HttpService;
-import org.nervos.neuron.service.httpservice.AppChainRpcService;
+import org.nervos.neuron.service.http.HttpUrls;
+import org.nervos.neuron.service.http.HttpService;
+import org.nervos.neuron.service.http.AppChainRpcService;
 import org.nervos.neuron.util.NetworkUtil;
 import org.nervos.neuron.util.db.DBAppUtil;
 import org.nervos.neuron.util.db.SharePrefUtil;
