+++ app/src/main/java/org/nervos/neuron/view/webview/JsInjectorClient.java
@@ -21,7 +21,7 @@ import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
 
-import org.nervos.neuron.service.httpservice.HttpUrls;
+import org.nervos.neuron.service.http.HttpUrls;
 import org.nervos.neuron.view.webview.item.Address;
 
 class JsInjectorClient {
